+++ packages/amplify-cli-core/src/utils/packageManager.ts
@@ -26,12 +26,10 @@ export interface PackageManager {
   getInstallArgs: (buildType: BuildType, resourceDir?: string) => string[];
 }
 
-const isWindows = process.platform === 'win32';
-
 class NpmPackageManager implements PackageManager {
   readonly packageManager = 'npm';
   readonly displayValue = 'NPM';
-  readonly executable = isWindows ? 'npm.cmd' : 'npm';
+  readonly executable = 'npm';
   readonly lockFile = 'package-lock.json';
 
   getRunScriptArgs = (scriptName: string) => ['run-script', scriptName];
@@ -41,7 +39,7 @@ class NpmPackageManager implements PackageManager {
 class YarnPackageManager implements PackageManager {
   readonly packageManager: PackageManagerType = 'yarn';
   readonly displayValue = 'Yarn';
-  readonly executable = 'yarn'; // Windows does not require `.cmd` extension to invoke yarn
+  readonly executable = 'yarn';
   readonly lockFile = 'yarn.lock';
   version?: SemVer;
 
@@ -77,7 +75,7 @@ class YarnPackageManager implements PackageManager {
 class PnpmPackageManager implements PackageManager {
   readonly packageManager: PackageManagerType = 'pnpm';
   readonly displayValue = 'PNPM';
-  readonly executable = isWindows ? 'pnpm.cmd' : 'pnpm';
+  readonly executable = 'pnpm';
   readonly lockFile = 'pnpm-lock.yaml';
 
   getRunScriptArgs = (scriptName: string) => [scriptName];