+++ packages/amplify-app/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.0.20](https://github.com/aws-amplify/amplify-cli/compare/@aws-amplify/amplify-app@5.0.19...@aws-amplify/amplify-app@5.0.20) (2023-10-05)
+
+**Note:** Version bump only for package @aws-amplify/amplify-app
+
+
+
+
+
+## [5.0.19](https://github.com/aws-amplify/amplify-cli/compare/@aws-amplify/amplify-app@5.0.18...@aws-amplify/amplify-app@5.0.19) (2023-10-03)
+
+**Note:** Version bump only for package @aws-amplify/amplify-app
+
+
+
+
+
+## [5.0.18](https://github.com/aws-amplify/amplify-cli/compare/@aws-amplify/amplify-app@5.0.17...@aws-amplify/amplify-app@5.0.18) (2023-09-22)
+
+**Note:** Version bump only for package @aws-amplify/amplify-app
+
+
+
+
+
+## [5.0.17](https://github.com/aws-amplify/amplify-cli/compare/@aws-amplify/amplify-app@5.0.16...@aws-amplify/amplify-app@5.0.17) (2023-08-31)
+
+**Note:** Version bump only for package @aws-amplify/amplify-app
+
+
+
+
+
 ## [5.0.16](https://github.com/aws-amplify/amplify-cli/compare/@aws-amplify/amplify-app@5.0.15...@aws-amplify/amplify-app@5.0.16) (2023-08-17)